backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
openai==1.40.1
httpx==0.27.2
uvicorn==0.30.6
waitress==2.1.2
PyYAML==6.0.2
pymilvus==2.4.4
marshmallow==3.21.3